rfc42: add credit based flow control (write direction)
Problem: rfc42 does not specify anything about channel flow control,
so it is difficult to design clients and servers that operate without
dropping data.

Add credit based flow control to the protocol in the write direction.

The client receives credit for writes to the subprocess via the
exec add-credit response.