rfc15: IMP input no longer uses stdin
Problem: the RFC states that the IMP takes its input on stdin
to avoid placing sensitive data on the command line, but stdin
is no longer used for this.

Now the IMP obtains its input by calling a helper program provided
by the instance instead of stdin.  The helper is run from the
unprivileged part of the IMP.

For now, just drop the incorrect detail which wasn't necessary
in that part of the text anyway.

See also: flux-framework/flux-security#163