rfc15: describe signal forwarding details
Problem: the IMP kill subcommand is briefly mentioned as the way
to signal guest processes, but this is inadequate in practice.

Now that the IMP lingers, just have it forward signals to the job
shell.  In addition, describe a surrogate signal that tells the
IMP to do its best to clean up the entire job container.

See also: flux-framework/flux-core#6011