rfc27: add 'free' key to hello response
Problem: when resources are partially allocated across a scheduler
reload, the HELLO protocol has no way to inform the scheduler
which resources of the job are allocated or freed.

Add an optional "free" idset key to the response.  If the key
is present, the ranks in the idset should be considered free.

See also: flux-framework/flux-core#6089