rfc11: support optional size key
Problem: It would be convenient to know the size of the data in
a val or valref object without decoding/reading the data.  However
there is no metadata for that.

Support an optional size key in 'val' and 'valref' objects.  It
allows us to know the size of the data without wasting cycles
decoding or reading the data.