mergify: remove explicit status checks
Problem: Explicit status checks are required in mergify conditions,
but these are already protected by branch protection rules on the
repo, so these conditions are redundant as well as difficult to
modify.

Drop the redundant conditions from `.mergify.yml`. Additionally,
move the remaining conditions under `pull_request_rules` as is done
in the other framework projects.