rfc42: add background flag to exec protocol
Problem: There is no way to run a subprocess in the background in the
RFC 42 exec protocol, i.e., allow the subprocess to continue running
as a child of the exec server without any client attached to receive
further responses in the RPC stream.

Add a background flag to the exec protocol. When set, the exec RPC is
initiated with a non-streaming request and the server replies with a
single response when the subprocess starts or fails. The subprocess
then continues running as a child of the server until it terminates
or the server shuts down.