rfc42: use RPC streaming flag instead of explicit background flag
Problem: The subprocess protocol defines a background flag to run
a process in the background. But this requires also dropping the
streaming flag from the RPC request, which may be redundant.

Drop the explicit background flag and instead run process in background
when the RPC is initiated with a non-streaming response.