rfc43: update job-list.list to streaming RPC
Problem: job-list.list RPC currently responds to requests with
a single large array of jobs.  This response can be extremely
large, and could lead to head-of-line blocking.

Update the job-list.list RPC to support streaming RPCs.  This will
allow jobs to be streamed back to the caller in smaller increments.