rfc43: update job-list.list to streaming RPC
Problem: job-list.list RPC currently responds to requests with
a single large array of jobs.  This response can be extremely
large, and could lead to head-of-line blocking.

Update the job-list.list RPC to be a streaming RPC.  It will instead
stream jobs back to the caller.

Allow backwards compatibility by allowing the older response to be
supported for non-streaming RPC requests.