rfc1: refine maintainer PR review process
Problem: RFC 1 says that maintainers must not make value judgements
on correct PRs, must merge them rapidly, and must express value
judgements as patches, which contradicts current project practice.

Instead of "rapidly", refer to RFC 48 which is intended to provide
remedies for "the maintainers are ignoring me!" problems.

Adjust the "value judgement" policy so that maintainers can continue to
ask contributors to rework PRs to increase their value and avoid needless
churn in the commit history.

This policy was likely added to C4.1 to solve a real problem with an
intransigent 0MQ maintainer.  We can always come up with a governance
remedy if Flux finds itself in a similar situation down the road, but
for now we seem to be doing OK.