rfc48: admin role in direction/conflict resolution
Problem: the current text explains the administrative aspects of the
admin role, but doesn't clarify the role in a way that would work for
some of the asks of projects in the HPSF charter for conflict resolution
or overall project direction.

Solution: Adapt some text from the HPCToolkit TSC
https://gitlab.com/hpctoolkit/governance/-/blob/main/GOVERNANCE.md#role
to clarify.